requirements.txt
numpy==1.26.4
pillow>=10.0.0
torch
torchvision
transformers>=4.40.0
huggingface_hub>=0.23.0
gradio>=5.0.0
